requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
aiohttp==3.9.3
pytest==8.0.2
black==24.2.0
isort==5.13.2
mypy==1.8.0 
flask==3.0.2